+++ tests/config/singletons-closure-not-return-type.php
@@ -5,9 +5,7 @@ declare(strict_types=1);
 return [
     'container' => [
         'singletons' => [
-            'closure-not-return-type' => static function () {
-                return new ArrayObject();
-            },
+            'closure-not-return-type' => static fn() => new ArrayObject(),
         ],
     ],
 ];